requirements.txt
fastapi==0.92.0
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
uvicorn==0.20.0
